requirements.txt
chainlit==2.6.0
langchain-core==0.3.67
langchain-openai==0.3.27
langgraph==0.5.1
langchain-mcp-adapters==0.1.8
toml
pyyaml